refactor: split postpro passes into separate files
